pyproject.toml
[project]
name = "calcohol"
version = "0.1.0"
description = "A calculator for selling drinks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml>=0.12.21",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.12.3",
]
